+++ jsolex-core/src/main/java/me/champeau/a4j/jsolex/expr/BuiltinFunction.java
@@ -78,6 +78,7 @@ public enum BuiltinFunction {
     STACK,
     SORT,
     VFLIP,
+    FILTER,
     VIDEO_DATETIME,
     WORKDIR(true);