+++ jsolex/src/main/resources/whats-new.md
@@ -11,6 +11,7 @@ Here are the new features in this version:
 
 - Enhanced reconstruction view which allows finding which frame is the source for a particular pixel (issue #386)
 - Improve spectral line identification in spectrum browser
+- Add `filter` ImageMath function, which allows filtering a list of images, for example to keep only these which are after a certain date
 
 ## Changes in 2.6.4